pyproject.toml
[project]
name = "microservice-template"
version = "0.1.0"
description = ""
authors = [
    {name = "andre",email = "andrepedro2004@hotmail.com"}
]
readme = "README.md"
requires-python = ">=3.12"
dependencies = [
    "fastapi (>=0.115.11,<0.116.0)",
    "grpcio (>=1.70.0,<2.0.0)",
    "grpcio-tools (>=1.70.0,<2.0.0)",
    "sqlmodel (>=0.0.23,<0.0.24)",
    "alembic (>=1.15.1,<2.0.0)",
    "sqlalchemy (>=2.0.38,<3.0.0)",
    "databases (>=0.9.0,<0.10.0)",
    "pymongo (>=4.11.2,<5.0.0)",
    "odmantic (>=1.0.2,<2.0.0)"
]

[tool.poetry]
#packages = [{include = "microservice_template", from = "src"}]


[tool.poetry.group.dev.dependencies]
uvicorn = "^0.34.0"
pytest = "^8.3.5"
pytest-asyncio = "^0.25.3"
mypy = "^1.15.0"
black = "^25.1.0"
isort = "^6.0.1"

[build-system]
requires = ["poetry-core>=2.0.0,<3.0.0"]
build-backend = "poetry.core.masonry.api"
